pyproject.toml
[project]
name = "py-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "kavlan-dev"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
]

[tool.poetry]
packages = [{include = "py_blog", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
